app/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0
pytest>=7.3.1
httpx>=0.24.0
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart
bcrypt>=4.0.1
requests>=2.31.0
httpx>=0.24.0
asyncio>=3.4.3
psycopg2-binary>=2.9.3  # For PostgreSQL support